requirements.txt
browser-use
langchain-anthropic
python-dotenv
